+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Models/AssetType.cs
@@ -1,0 +1,60 @@
+//    #[license]
+//    SmartsheetClient SDK for C#
+//    %%
+//    Copyright (C) 2014 SmartsheetClient
+//    %%
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//        
+//            http://www.apache.org/licenses/LICENSE-2.0
+//        
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+//    %[license]
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace Smartsheet.Api
+{
+    /// <summary>
+    /// Represents the type of asset for sharing operations.
+    /// </summary>
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum AssetType
+    {
+        /// <summary>
+        /// Sheet asset type
+        /// </summary>
+        SHEET,
+
+        /// <summary>
+        /// Report asset type
+        /// </summary>
+        REPORT,
+
+        /// <summary>
+        /// Sight asset type
+        /// </summary>
+        SIGHT,
+
+        /// <summary>
+        /// Workspace asset type
+        /// </summary>
+        WORKSPACE,
+
+        /// <summary>
+        /// Collection asset type
+        /// </summary>
+        COLLECTION,
+
+        /// <summary>
+        /// File asset type
+        /// </summary>
+        FILE
+    }
+}